+++ back/jest.config.js
@@ -5,7 +5,6 @@ module.exports = {
   transform: {
     "^.+\\.[tj]sx?$": ["ts-jest", { tsconfig: "<rootDir>/tsconfig.spec.json" }]
   },
-  testTimeout: 60000,
   testPathIgnorePatterns: ["<rootDir>/dist/"],
   testMatch: ["**/__tests__/**/?(*.)+(spec|test).[jt]s?(x)"],
   reporters: ["default", "github-actions"]
