+++ back/src/bsvhu/resolvers/mutations/__tests__/duplicateBsvhu.integration.ts
@@ -1,5 +1,5 @@
 import { gql } from "graphql-tag";
-import { xDaysAgo } from "../../../../commands/onboarding.helpers";
+import { xDaysAgo } from "../../../../utils";
 import { resetDatabase } from "../../../../../integration-tests/helper";
 import {
   companyFactory,
