+++ back/src/forms/resolvers/mutations/__tests__/duplicateForm.integration.ts
@@ -11,7 +11,7 @@ import {
   userWithCompanyFactory
 } from "../../../../__tests__/factories";
 import makeClient from "../../../../__tests__/testClient";
-import { xDaysAgo } from "../../../../commands/onboarding.helpers";
+import { xDaysAgo } from "../../../../utils";
 import {
   getFirstTransporter,
   getFirstTransporterSync
