+++ back/src/users/resolvers/mutations/__tests__/createPasswordResetRequest.integration.ts
@@ -4,8 +4,7 @@ import { resetDatabase } from "../../../../../integration-tests/helper";
 import prisma from "../../../../prisma";
 import { Mutation } from "../../../../generated/graphql/types";
 import { sendMail } from "../../../../mailer/mailing";
-import { createPasswordResetRequest } from "../../../../mailer/templates";
-import { renderMail } from "../../../../mailer/templates/renderers";
+import { renderMail, createPasswordResetRequest } from "@td/mail";
 import { addMinutes } from "date-fns";
 import { setCaptchaToken } from "../../../../common/redis/captcha";
 import { gql } from "graphql-tag";
