+++ back/src/utils.ts
@@ -69,6 +69,20 @@ export function daysBetween(date1: Date, date2: Date): number {
   return numberOfdays;
 }
 
+/**
+ * Compute a past date relative to baseDate
+ *
+ * @param baseDate Date
+ * @param daysAgo Integer
+ * @return a date at 00:00:00
+ */
+export const xDaysAgo = (baseDate: Date, daysAgo: number): Date => {
+  const clonedDate = new Date(baseDate.getTime()); // avoid mutating baseDate
+  clonedDate.setDate(clonedDate.getDate() - daysAgo);
+
+  return new Date(clonedDate.toDateString());
+};
+
 export function sanitizeEmail(email: string): string {
   return email.toLowerCase().trim();
 }
