--- libs/back/mail/src/lib/libs/back/mail.spec.ts
@@ -1,7 +1,0 @@
-import { libsBackMail } from "./libs/back/mail";
-
-describe("libsBackMail", () => {
-  it("should work", () => {
-    expect(libsBackMail()).toEqual("libs/back/mail");
-  });
-});
