--- libs/back/mail/src/lib/libs/back/mail.ts
@@ -1,3 +1,0 @@
-export function libsBackMail(): string {
-  return "libs/back/mail";
-}
